requirements.txt
PyJWT==2.1.0
fastapi==0.63.0
passlib==1.7.4
bcrypt==3.2.0
motor==2.4.0
boto3==1.17.72
aiofiles==0.6.0
uvicorn==0.13.4
